ADEN-5269 Rename file
